pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docverificator"
version = "0.1.0"
description = "Document OCR + field extraction + authenticity-oriented validation for IDs and statements."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26.0,<2.0",
  "opencv-python>=4.6.0,<4.7.0",
  "Pillow>=10.3.0",
  "python-dateutil>=2.9.0",
  "PyMuPDF>=1.24.0",
  "rapidfuzz>=3.9.0",
]

[project.optional-dependencies]
ocr-paddle = ["paddleocr==2.7.3", "paddlepaddle==2.6.2"]
ocr-easy = ["easyocr>=1.7.1"]
heic = ["pillow-heif>=0.16.0"]
dev = ["pytest>=8.2.0"]

[project.scripts]
docverificator = "docverificator.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
